Refactoring editing by extract method
The original methods relied on program state for in/out points and
what clip to edit into the sequence (and still do). The extracted
methods take that data as arguments instead.

The rationale for the refactoring is to break out a method that can be
used for loading of sequences, while the original method is used for
interactive editing.

FIX: MainWindow::program_insert was refactored into two methods.

FIX: MainWindow::program_overwrite was refactored into two methods.