Refactoring MainWindow into a singleton
The rationale for the refactoring is to make the application instance
(at this moment of the MainWindow class) globaly accessable, so that,
for instance, a sequence loader can call the applications/documents
public methods.

This is part of the work underway to refactor the application into
proper MVC.

ADD: Added a factory for the MainWindow that will return a global
     instance of the MainWindow if it has already been instanced.

TODO: Refactor the factory to be absract.