Implementing document loading and enabling it in the GUI
Enabled the "File/Open" menu item so that Iola XML documents can be
loaded in the user interface.

Created a parser for loading a minimal Iola XML document. The format
is partly modeled on FCP XML and use the same element names but has a
slightly different structure.

There are two types of elements in the parser: containers and
terminals. The former contain other elements while the latter simply
enclose a single value.

An Iola XML document will update the application when restore() is
called. Calling restore() on an element will also recursively call it
on all its children.