Implementing document saving and enabling it in the GUI
Enabled the "File/Save" and "File/Save As" menu items so that Iola XML
documents can be saved in the user interface.

An Iola XML document will update itself when store() is called.
Calling restore() on an element will also recursively call it on all
its children. The result can then be written out to a stream by
calling xml().