Renaming src/ dir to iola/
Renaming the src/ directory to iola/ and making adding the top source
directory to include directories so that headers can be included using
angled brackets, in a sensible way, instead of quote marks. E.g.:

  #include <iola/dom/root.h>

Thus avoiding relative paths that might break and must be updated if
source files get moved around.

The directory structure will also match the namespace, so that the
included class in the example above will be accessed as:

  iola::dom::root()