Refactoring application into MVC design by extract class
The original class MainWindow acted both as the application window for
the GUI (with SourceMonitor and ProgramMonitor as view/controller for
source and program) and as the model. The extracted class project take
over the role as model and leave MainWindow as view/controller.

The original class MainWindow also acted as application instance
singleton. The extracted class application_implementation (an
implementation of the iapplication interface) takes over that role.

The rationale for the refactoring is to reorganize the code and break
it up in logical modules and roles. The new module domain contain all
domain logic (i.e. the model in MVC). The new module application
contain the application singleton. While the gui module contain the
view/controller logic for the graphical user interface.

ADD: New module application.

ADD: New module domain.