Refactoring producer change signals by move method
Moved the producer listener from the monitor classes into the project
class. The listener was modified to emit a new signal that the monitor
objets can connect to instead to recieve updates on when their
respective producer has changed.

The rationale for the refactoring is to isolate and make the gui and
the project classes less coupled. With the motivation that the source
and program objects are fields in the project class, it is the project
class that is supposed to access them directly. The goal is to
deprecate the need for the get_source and get_program methods.