Refactoring source/program access by encapsulate field
Added getters to the project class for start/end frame numbers. Other
classes are supposed to use them instead of getting the numbers
directly from the source or program object.

The rationale for the refactoring is to isolate and make the gui and
the project classes less coupled. With the motivation that the source
and program objects are fields in the project class, it is the project
class that is supposed to access them directly. The goal is to
deprecate the need for the get_source and get_program methods.

ADD: Methods to get source start/end frame number.

ADD: Methods to get program start/end frame number.