Refactoring source/program access by encapsulate field
Added consumer connect method to the project so that the gui can
register consumers with the project instead of directly with the
source/program objects.

The rationale for the refactoring is to isolate and make the gui and
the project classes less coupled. With the motivation that the source
and program objects are fields in the project class, it is the project
class that is supposed to access them directly. The goal is to
deprecate the need for the get_source and get_program methods.

ADD: Method to register a source consumer with the project.

ADD: Method to register a program consumer with the project.