Removing deprecated source/program getter methods
The rationale for the refactoring is to isolate and make the gui and
the project classes less coupled. With the motivation that the source
and program objects are fields in the project class, it is the project
class that is supposed to access them directly. The goal to deprecate
the need for the get_source and get_program methods has now been
accomplished.

REM: Method to get the source producer object.

REM: Method to get the program producer object.