Fixing deadlock bug when consumer is stopping.
Apparently a deadlock bug sometimes reared its dirty head when the
consumer thread was being stopped. The fault turned out to be the
firing of the consumer-frame-show event that made the consumer thread
execute a callback function that would hang in a deadlock.

The solution is to remove the Fl::lock() from the callback. Note that
this simply just avoids the situation. I do not yet understand how the
deadlock is created in the first place (was more interested in solving
the problem and move on rather than analyzing it with Helgrind).

The Fl::check() call was also removed from the callback function and
replaced with Fl::wait(0) in the main loop. Calling Fl::check() in the
callback essentially mean that the GUI is updated by the consumer
thread instead of the main loop in the main thread. Which is stupid.