Improving XML output by adding newlines.
Improving the XML output by adding newlines. Only terminal elements
have both their start-tag, content and end-tag on the same line.

The rationale is to make the XML output easier to scan by eye (altough
it's recomended to use an unstyled XML viewer such as a web-browser)
and improve the formats suitability for revision control (as diff work
on a line basis).