Removing image jitter on playback. Closes gh-23
The OpenGL consumer could not currently handle the correct ITU-R
Rec. BT.601 pixel aspect ratios without having the image jittering
unstable on playback.

Turned out that the solution was simple: The consumer property
"rescale" defaults to "bilinear" and was passed to all scaling MLT
filters. Setting it to "none" removed the jittering.

There's no rationale for having MLT scaling the image for us as the
OpenGL consumer is quite capable in that area.