Improving the robustness of the XML parser. Closes gh-35
If an element did not exist in the DOM, the element factory returned a
`NULL` pointer to the parser. This causes problem with popping
elements from the context stack in `endElement` as `NULL` pointer
elements (for obvious reasons) is never pushed on the stack in
`startElement`.

The solution is to never have the element factory return a `NULL`
pointer but a dummy element instead.